fixup trace logic to be shareable
